Added Click Event for New Game
Created a new GameObject "MenueEvents" as a container for  scripts for a menue level respectively. Added the child "EventsMainMenue" of  "MenueEvents" with the click function "StartNewGame" to the button New Game. Now when clicked scene changes to a certain scene. Scene is assigned by giving the name of it as parameter to the click function. Created scene "Level".