Prepared Menue for interaction with stats of  Player
Change Name from StartNewGame to ChangeScene to indicate that the can load any scene
Added Event System to the prefab PlayerCamera.
Created a script "ReloadScene" to reset the current scene.
Created a script "UpdateStatsText".
Cretaed a script "ManageGameOverScreen"
Added component "UpdateStatsText" to the TextHealth to prefab "PlayerCamera".
Added component "UpdateStatsText" to the TextScore to prefab "PlayerCamera"
Move object "GameOverScreen" from prefab "PlayerCamera" and made it into a prefab.
Now GameOverScreen is created at runtime if death of player as event is fired.